requirements.txt
airweave-sdk
streamlit
pandas
groq
